Moved spew to common namespace + fixes references + meta files
