+++ client/Packages/com.beamable/Common/Runtime/Spew.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1221bf503e0046d0b6ecad86c963342f
+guid: 593ab8def454170419105398dc498e4b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
