+++ client/Packages/com.beamable/Common/Runtime/Spew/SpewLoggerAttribute.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Diagnostics;
+
+namespace Beamable.Common.Spew
+{
+	/// <summary>
+	/// Conditional attribute to add Spew.
+	/// </summary>
+	[Conditional("UNITY_EDITOR")]
+	[AttributeUsage(AttributeTargets.Class)]
+	public sealed class SpewLoggerAttribute : Attribute { }
+}
