+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/Chat/ChatProvider.cs
@@ -1,7 +1,7 @@
 using Beamable.Common;
 using Beamable.Common.Api.Notifications;
 using Beamable.Common.Dependencies;
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using System;
 using System.Collections.Generic;
 using UnityEngine;
