+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/Chat/PubNubRoom.cs
@@ -4,7 +4,7 @@ using Beamable.Common;
 using Beamable.Common.Api.Notifications;
 using Beamable.Common.Dependencies;
 using Beamable.Serialization;
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using System.Collections.Generic;
 using Debug = UnityEngine.Debug;
