+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/Notification/Internal/InGameNotification.cs
@@ -1,4 +1,4 @@
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using System;
 
 namespace Beamable.Api.Notification.Internal
