+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/Notification/Internal/MultiChannelSubscriptionHandler.cs
@@ -1,4 +1,4 @@
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using PubNubMessaging.Core;
 using System;
