+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/Notification/NotificationService.cs
@@ -1,6 +1,6 @@
 using Beamable.Api.Notification.Internal;
 using Beamable.Common.Api.Notifications;
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using System;
 using System.Collections.Generic;
 using UnityEngine;
