+++ client/Packages/com.beamable/Runtime/Core/Platform/SDK/PlatformSubscribable.cs
@@ -5,7 +5,7 @@ using Beamable.Common.Api;
 using Beamable.Common.Api.Notifications;
 using Beamable.Common.Dependencies;
 using Beamable.Coroutines;
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using System;
 using System.Collections;
 using System.Collections.Generic;
