+++ client/Packages/com.beamable/Runtime/Core/Pooling/Editor.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e438a083ef7cd0542a4095e71e13cdc4
+guid: c7d02dd4e10947e439afc1f3da83984f
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
