+++ client/Packages/com.beamable/Runtime/Core/Pooling/Editor/Beamable.Pooling.Editor.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 895213f17c00b469d947179e147ea5cf
-folderAsset: yes
-DefaultImporter:
+guid: aeb22b93e292d004faeafb3f4ee17868
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
