+++ client/Packages/com.beamable/Runtime/Core/Serialization/JsonSerializable/LoadStream.cs
@@ -1,5 +1,5 @@
 using Beamable.Pooling;
-using Beamable.Spew;
+using Beamable.Common.Spew;
 using System;
 using System.Collections;
 using System.Collections.Generic;
