--- client/Packages/com.beamable/Runtime/Core/Spew/Beamable.Spew.asmdef
@@ -1,14 +1,0 @@
-{
-    "name": "Beamable.Spew",
-    "references": [
-        "Unity.Beamable.Runtime.Common"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": []
-}
