+++ client/Packages/com.beamable/Runtime/Unity.Beamable.asmdef
@@ -7,7 +7,6 @@
         "Beamable.Platform",
         "Beamable.Service",
         "Beamable.Pooling",
-        "Beamable.Spew",
         "Beamable.JsonSerializable",
         "Beamable.SmallerJSON",
         "Beamable.Runtime.Input",
